+++ apps/chronicles/e2e/jest.config.js
@@ -7,8 +7,8 @@ module.exports = {
     maxWorkers: 1,
     setupFiles: ["<rootDir>/e2e/helpers.js"],
     globalSetup: "detox/runners/jest/globalSetup",
-     globalTeardown: "detox/runners/jest/globalTeardown",
-     reporters: ["detox/runners/jest/reporter"],
-     testEnvironment: "detox/runners/jest/testEnvironment",
+    globalTeardown: "detox/runners/jest/globalTeardown",
+    reporters: ["detox/runners/jest/reporter"],
+    testEnvironment: "detox/runners/jest/testEnvironment",
     verbose: true,
 };