+++ apps/chronicles/metro.config.js
@@ -3,6 +3,7 @@
 // Learn more https://docs.expo.io/guides/customizing-metro
 
 const { getDefaultConfig } = require("expo/metro-config");
+const { wrapWithReanimatedMetroConfig } = require("react-native-reanimated/metro-config");
 const path = require("path");
 
 // Find the project and workspace directories
@@ -20,4 +21,4 @@ config.resolver.nodeModulesPaths = [
     path.resolve(monorepoRoot, "node_modules"),
 ];
 
-module.exports = config;
+module.exports = wrapWithReanimatedMetroConfig(config);