+++ packages/components/__tests__/Typography.spec.tsx
@@ -1,13 +1,13 @@
-import React, { LegacyRef, MutableRefObject, RefObject, createRef } from "react";
-import { act, render } from "@testing-library/react-native";
-import { Typography } from "../src/components/Typography";
+import { act } from "@testing-library/react-native";
+import React, { createRef } from "react";
 import { Text } from "react-native";
+import { Typography } from "../src/components/Typography";
+import { renderWithTheme } from "../test-utils";
 
 describe("Typography", () => {
     it("Should forward the ref to the Text element", () => {
         const ref = createRef<Text>();
-        render(<Typography ref={ref}>Typography</Typography>);
-        //@ts-expect-error current can be null, but we don't care
+        renderWithTheme(<Typography ref={ref}>Typography</Typography>);
         const blur = jest.spyOn(ref.current, "blur");
         act(() => {
             //Text component has a blur method. Typography does not.