+++ packages/core/src/store/index.ts
@@ -1,3 +1,3 @@
 export * from "./demo-mode";
 export * from "./language";
-export {useMadConfig, setEnvironment, getConfig } from "./mad-config";
+export { useMadConfig, setEnvironment, getConfig } from "./mad-config";