+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @equinor/mad-auth
 
+## 0.2.9
+
+### Patch Changes
+
+-   Updated dependencies [7ff2d25]
+-   Updated dependencies [75df020]
+-   Updated dependencies [75df020]
+-   Updated dependencies [2ce9392]
+-   Updated dependencies [75df020]
+    -   @equinor/mad-components@0.16.0
+
 ## 0.2.8
 
 ### Patch Changes